feat: move connectionState onto top level room object
